Bug in load balancing is fixed.

+++ test/sharding/integration/mini_cluster/init-cluster.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 . ./cluster_config.sh
 
-###################
+##############################################################
 # Files that must be copied to each machine in install process
 __INSTALL_FILES_SCRIPTS=( run_cluster_command \
                   init-group.sh \
@@ -9,7 +9,7 @@ __INSTALL_FILES_SCRIPTS=( run_cluster_command \
                   env-util.sh \
                   generateConfigFile.py \
                   init-env.sh \
-                  git-operation.sh \
+                  branch-op.sh \
                   run-group.sh \
                   generateFrontendConfig.py \
                   frontend \
@@ -18,37 +18,68 @@ __INSTALL_FILES_DATA=( $__LICENSE_FILE \
                   $__STOP_WORDS \
                   $__PROTECTED_WORDS \
                   conf-template.xml)
-###################
-
-if [[ $@ -gt 1 ]] ; then
-   refreshFlag=$1
-else
-   refreshFlag=0
-fi
-
-echo "Copying needed scripts on all clients ..."
+###############################################################
+formatAll=0
+forceUpdate=0
+for var in "$@"
+do
+   case $var in
+   -a)
+      formatAll=1
+      continue 
+   ;;
+   -na)
+      formatAll=0
+      continue
+   ;;
+   -f)
+      forceUpdate=1
+      continue
+   ;;
+   -nf)
+      forceUpdate=0
+      continue
+   ;;
+   *)
+      echo "Option not recognized.\nUsage : bash init-cluster.sh [-a{format all}] [-f{force update all}]"
+      exit
+   ;;
+   esac
+done
+#echo "Read args are $1 and $2"
+#echo "Args are $formatAll and $forceUpdate ."
+echo "Going to copy modified(new) scripts on all clients ..."
 for i in "${__GROUPS[@]}"
 do
-   if [[ $refreshFlag ]] ; then
+   if [[ $formatAll -eq 1 ]] ; then
       ssh ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]} rm -r ${__SRCH2_HOMES[$i]}
    fi
-   #echo "${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]} mkdir ${__SRCH2_HOMES[$i]}"
    ssh ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]} mkdir ${__SRCH2_HOMES[$i]}
    ssh ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]} mkdir ${__SRCH2_HOMES[$i]}/$__BIN_DIR_NAME
    for f in "${__INSTALL_FILES_SCRIPTS[@]}"
    do
+      ###### Only upload the modified scripts  ######
+      if [ -z `find ./$__BIN_DIR_NAME/$f -newer .lastChangeTime` ] && [ $forceUpdate -eq 0 ] ; then
+         continue
+      fi
+      ###### Transferring the file ######
+      #echo "File ./$__BIN_DIR_NAME/$f has modified." 
       if [[ -d ./$__BIN_DIR_NAME/$f ]]; then
           scp -r ./$__BIN_DIR_NAME/$f ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]}:${__SRCH2_HOMES[$i]}/$__BIN_DIR_NAME
       else
           scp ./$__BIN_DIR_NAME/$f ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]}:${__SRCH2_HOMES[$i]}/$__BIN_DIR_NAME
       fi
-      #ssh ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]} $(echo "'chmod 777 ${__SRCH2_HOMES[$i]}/run_cluster_command'")
    done
    for f in "${__INSTALL_FILES_DATA[@]}"
    do
+      ###### Only upload the modified scripts  ######
+      if [ -z `find ./$f -newer .lastChangeTime` ] && [ $forceUpdate -eq 0 ] ; then
+         continue
+      fi
+      ###### Transferring the file ######
+      #echo "File ./$f has modified." 
       scp ./$f ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]}:${__SRCH2_HOMES[$i]}
-      #ssh ${__LOGIN_USERS[$i]}@${__IP_ADDRESSES[$i]} $(echo "'chmod 777 ${__SRCH2_HOMES[$i]}/run_cluster_command'")
    done
-
 done
+touch -m .lastChangeTime
 echo ${__IP_ADDRESSES[*]}
